+++ fastNLP/core/__init__.py
@@ -2,7 +2,7 @@ from .batch import Batch
 from .dataset import DataSet
 from .fieldarray import FieldArray
 from .instance import Instance
-from .losses import Loss
+from .losses import LossFromTorch
 from .optimizer import Optimizer
 from .sampler import SequentialSampler, BucketSampler, RandomSampler, BaseSampler
 from .tester import Tester
