+++ reproduction/LSTM+self_attention_sentiment_analysis/main.py
@@ -2,8 +2,8 @@ import torch.nn.functional as F
 
 from fastNLP.core.trainer import ClassificationTrainer
 from fastNLP.core.utils import ClassPreprocess as Preprocess
-from fastNLP.io.config_loader import ConfigLoader
-from fastNLP.io.config_loader import ConfigSection
+from fastNLP.io.config_io import ConfigLoader
+from fastNLP.io.config_io import ConfigSection
 from fastNLP.io.dataset_loader import ClassDataSetLoader as Dataset_loader
 from fastNLP.models.base_model import BaseModel
 from fastNLP.modules.aggregator.self_attention import SelfAttention
