+++ setup.py
@@ -12,12 +12,12 @@ with open('requirements.txt', encoding='utf-8') as f:
     reqs = f.read()
 
 setup(
-    name='fastNLP',
+    name='FastNLP',
     version='0.1.1',
     description='fastNLP: Deep Learning Toolkit for NLP, developed by Fudan FastNLP Team',
     long_description=readme,
     license=license,
-    author='fudanNLP',
+    author='FudanNLP',
     python_requires='>=3.5',
     packages=find_packages(),
     install_requires=reqs.strip().split('\n'),
